entrypoint: Disable binlog while loading tzinfo
When running mysqlsh clusterAddInstance to create a cluster the
following warnings are added to the output of mysqlsh

This change turns off SQL_BIN_LOG so that the GTID is not generated by
the initial tzinfo insert statements.

```
WARNING: A GTID set check of the MySQL instance at 'hostname.svc.cluster.local:3306' determined that it contains transactions that do not originate from the cluster, which must be discarded before it can join the cluster.
hostname.svc.cluster.local:3306 has the following errant GTIDs that do not exist in the cluster:
c8ee89a9-840a-11eb-8a25-0242ac11000b:1-5
```

Co-authored-by: David Sharp <dsharp@vmware.com>
Co-authored-by: Colin Shield <cshield@vmware.com>